Remove ambiguous warning about garbage output
When QOA was first released I wasn't exactly sure if this compression
scheme will stick; and people were invited to experiment with it. So
this warning was meant for developers who'd like to modify the code.
E.g.: changing the shift values in the encoder and neglecting to change
them in the decoder will produce some very bad results.